Code now cleaned.
All the code is now cleaned. Doing a test revealed that the entry editor does not save changes to an entry, and because of this it is unknown if the entry ID's will update upon an entry deletion. Entry deletion does not work either. This all indicates an issue with the new method of searching the database for a matching id...