Fixed deletion and ID updating
Seperation between the entry ID for the user and its primary key for the database is complete. When an entry is deleted, the other user based entry IDs are updated, so even though holes are made in the databse primary key list, user entered IDs still start at 1 and go forever with no holes.