Added final polish, version 1.0 is complete
Added some final polish, removed the settings button as it was determined that no settings needed to be changed. Added a prompt in the lsit view to tell the user that there are no entries, and made it so they cannot tap on that prompt.

NOTE: I was very lazy when removing the settings page, I just deleted the UI button, all the code is still there, perhaps as a seed for the future?